[FEATURE] Add `#limit` method to `Relation`
Returns a new relation with the `:limit` option added to its dataset.

The `#using` option is not restricted in any way. You can change
every dataset attribute with that option. This is because dataset
session is not preinitialized any longer. Every update costs nothing
in terms of efficiency.

Methods `#limit`, `#offset` and `#where` are built on top of that
universal `#using`.

== Added

* ROM::Kafka::Relation#limit